Extracted funcs related to machine files and cli args
